pyproject.toml
[project]
name = "bond-ai"
version = "0.0.1"
description = "BOND: Biomedical Ontology Normalization and Disambiguation"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Pankaj Rajdeo" }]
license = { text = "MIT" }

dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anthropic==0.64.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "chardet==5.2.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cohere==5.17.0",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "et_xmlfile==2.0.0",
    "faiss-cpu==1.8.0",
    "fastapi==0.116.1",
    "fastavro==1.12.0",
    "fastobo==0.13.0",
    "filelock==3.19.1",
    "filetype==1.2.0",
    "fonttools==4.59.1",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "grpcio==1.74.0",
    "grpcio-status==1.74.0",
    "h11==0.16.0",
    "hf-xet==1.1.7",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "importlib_metadata==8.7.0",
    "Jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "kiwisolver==1.4.9",
    "langchain==0.3.27",
    "langchain-anthropic==0.3.18",
    "langchain-cohere==0.4.5",
    "langchain-community==0.3.27",
    "langchain-core==0.3.74",
    "langchain-google-genai==2.1.9",
    "langchain-huggingface==0.3.1",
    "langchain-openai==0.3.30",
    "langchain-text-splitters==0.3.9",
    "langsmith==0.4.14",
    "litellm==1.75.8",
    "MarkupSafe==3.0.2",
    "marshmallow==3.26.1",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "mypy_extensions==1.1.0",
    "networkx==3.5",
    "numpy==1.26.4",
    "openai==1.99.9",
    "openpyxl==3.1.5",
    "orjson==3.11.2",
    "packaging==25.0",
    "pandas==2.3.2",
    "pillow==11.3.0",
    "pronto==2.7.0",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.32.0",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "pydantic_core==2.33.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "rdflib==7.1.4",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.27.0",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.1",
    "scipy==1.16.1",
    "sentence-transformers==5.1.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.43",
    "starlette==0.47.2",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.11.0",
    "tokenizers==0.21.4",
    "torch==2.2.2",
    "tqdm==4.67.1",
    "transformers==4.55.2",
    "types-PyYAML==6.0.12.20250809",
    "types-requests==2.32.4.20250809",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.23.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0", "ruff>=0.4.0", "black>=24.4.2"]

[project.scripts]
bond-query = "bond.cli:main"
bond-serve = "bond.server:main"
bond-build-faiss = "scripts.build_faiss_from_sqlite:main"
bond-generate-sqlite = "scripts.ontology_sqlite_generator:main"
bond-ingest-umls = "scripts.ingest_umls_full_pipeline:main"
bond-ingest-lungmap = "scripts.lungmap_provisional_ingest:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["bond*", "scripts*"]
